qtbase: update to latest (ie final) revision in upstream 5.4 branch
The previous revision was somewhere just after the 5.4.2 release tag
but a few revisions before the final commit in the qtbase 5.4 branch.

Updating to the final commit in the qtbase 5.4 branch adds:

  4a1e5db Better handling of invalid font tables
  eae0cb0 Network: Fix up previous corruption patch
  d82d5b1 Check for integer overflows in places where qAllocMore is used
  3401440 quote QUuid values
  d1ce295 Make QStyle::itemPixmapRect handle devicePixelRatio

Update by adding patches (instead of updating SRCREV) to avoid
needing to update existing git clones which users already have in
downloads.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>